Create mock.js file, update mock function
