+++ package.json
@@ -3,7 +3,7 @@
   "description": "",
   "private": true,
   "scripts": {
-    "test": "jest",
+    "test": "jest --coverage",
     "hint": "npx hint .",
     "hintf": "npx hint --fix .",
     "lint": "npx eslint .",
@@ -49,5 +49,11 @@
   "dependencies": {
     "lodash": "^4.17.21"
   },
-  "main": "webpack.config.js"
+  "main": "webpack.config.js",
+  "jest": {
+    "collectCoverage": true,
+    "CoverageReporters": [
+      "html"
+    ]
+  }
 }
